Corrige a exibição dos valores e quantidades

+++ database/migrations/2022_01_27_004456_create_products_table.php
@@ -17,7 +17,7 @@ class CreateProductsTable extends Migration
             $table->integer('id', true);
             $table->string('name');
             $table->text('description');
-            $table->double('price')->nullable();
+            $table->integer('price')->nullable();
             $table->string('image')->nullable();
             $table->timestamps();
             $table->softDeletes();
